🌱 Split feedback.go and gitops.go into domain-specific files
Split two large handler monoliths in pkg/api/handlers/:

feedback.go (2563 lines → 3 files, original deleted):
- feedback_config.go: Types, constants, constructor, utility
  functions (parsePageParams, extractHost, LoadFeedbackConfig) (343)
- feedback_requests.go: Feature request CRUD, listing, queue
  items, notifications, feedback submission (1312)
- feedback_github.go: GitHub webhook handling, issue/PR operations,
  screenshot uploads, PR comments, signature verification (939)

gitops.go (2467 lines → 552 + 3 files):
- gitops.go retains: types, constructor, drift cache, Helm list,
  Kustomizations (552)
- gitops_operators.go: Operator list/stream, subscriptions (788)
- gitops_drift.go: Drift detection (MCP + kubectl), sync, diff
  parsing, validators, git clone (645)
- gitops_argo.go: Argo CD applications, health/sync summaries,
  Helm history/values (460)

All methods remain on their respective handler structs in the same
 zero interface or behavior changes.package

Fixes #10419
